// src/interfaces/clienteInterface.ts
export interface IClienteAttributes {
    id_cliente: number;
    activo: boolean;
    contrato: string;
    días_front: number;
    eliminado: boolean;
    email: string;
    fecha_alta: Date;
    fecha_baja: Date;
    nombre: string;
    numero_telefono: string;
    pais: string;
    plan: string;
  }

  